crash: increase HTMLCollection's RC from u8 to u32

+++ src/browser/webapi/collections/HTMLCollection.zig
@@ -58,7 +58,7 @@ _data: union(Mode) {
     form: NodeLive(.form),
     empty: void,
 },
-_rc: lp.RC(u8) = .{},
+_rc: lp.RC(u32) = .{},
 
 pub fn deinit(self: *HTMLCollection, page: *Page) void {
     page.factory.destroy(self);
